LibWeb: Move gradient color stops expansion to recording phase
Expand color stops during display list recording rather than playback.
Recording happens once but the display list may be executed many times,
so doing this work at record time is more efficient.
