+++ Libraries/LibWeb/Painting/GradientPainting.cpp
@@ -5,6 +5,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
+#include <AK/Array.h>
 #include <AK/Math.h>
 #include <LibGfx/Gradients.h>
 #include <LibWeb/CSS/CalculationResolutionContext.h>
@@ -19,6 +20,110 @@
 #include <LibWeb/Painting/GradientPainting.h>
 
 namespace Web::Painting {
+
+static ColorStopList replace_transition_hints_with_normal_color_stops(ColorStopList const& color_stop_list)
+{
+    ColorStopList stops_with_replaced_transition_hints;
+
+    auto const& first_color_stop = color_stop_list.first();
+    // First color stop in the list should never have transition hint value
+    VERIFY(!first_color_stop.transition_hint.has_value());
+    stops_with_replaced_transition_hints.empend(first_color_stop.color, first_color_stop.position);
+
+    // This loop replaces transition hints with five regular points, calculated using the
+    // formula defined in the spec. After rendering using linear interpolation, this will
+    // produce a result close enough to that obtained if the color of each point were calculated
+    // using the non-linear formula from the spec.
+    for (size_t i = 1; i < color_stop_list.size(); i++) {
+        auto const& color_stop = color_stop_list[i];
+        if (!color_stop.transition_hint.has_value()) {
+            stops_with_replaced_transition_hints.empend(color_stop.color, color_stop.position);
+            continue;
+        }
+
+        auto const& previous_color_stop = color_stop_list[i - 1];
+        auto const& next_color_stop = color_stop_list[i];
+
+        auto distance_between_stops = next_color_stop.position - previous_color_stop.position;
+        auto transition_hint = color_stop.transition_hint.value();
+
+        Array transition_hint_relative_sampling_positions {
+            transition_hint * 0.33f,
+            transition_hint * 0.66f,
+            transition_hint,
+            transition_hint + (1.f - transition_hint) * 0.33f,
+            transition_hint + (1.f - transition_hint) * 0.66f,
+        };
+
+        for (auto const& transition_hint_relative_sampling_position : transition_hint_relative_sampling_positions) {
+            auto position = previous_color_stop.position + transition_hint_relative_sampling_position * distance_between_stops;
+            auto value = Gfx::color_stop_step(previous_color_stop, next_color_stop, position);
+            auto color = previous_color_stop.color.interpolate(next_color_stop.color, value);
+            stops_with_replaced_transition_hints.empend(color, position);
+        }
+
+        stops_with_replaced_transition_hints.empend(color_stop.color, color_stop.position);
+    }
+
+    return stops_with_replaced_transition_hints;
+}
+
+static ColorStopList expand_repeat_length(ColorStopList const& color_stop_list, float repeat_length)
+{
+    // https://drafts.csswg.org/css-images/#repeating-gradients
+    // When rendered, however, the color-stops are repeated infinitely in both directions, with their
+    // positions shifted by multiples of the difference between the last specified color-stop's position
+    // and the first specified color-stop's position. For example, repeating-linear-gradient(red 10px, blue 50px)
+    // is equivalent to linear-gradient(..., red -30px, blue 10px, red 10px, blue 50px, red 50px, blue 90px, ...).
+
+    auto first_stop_position = color_stop_list.first().position;
+    int const negative_repeat_count = AK::ceil(first_stop_position / repeat_length);
+    int const positive_repeat_count = AK::ceil((1.0f - first_stop_position) / repeat_length);
+
+    ColorStopList color_stop_list_with_expanded_repeat = color_stop_list;
+
+    auto get_color_between_stops = [](float position, auto const& current_stop, auto const& previous_stop) {
+        auto distance_between_stops = current_stop.position - previous_stop.position;
+        auto percentage = (position - previous_stop.position) / distance_between_stops;
+        return previous_stop.color.interpolate(current_stop.color, percentage);
+    };
+
+    for (auto repeat_count = 1; repeat_count <= negative_repeat_count; repeat_count++) {
+        for (auto stop : color_stop_list.in_reverse()) {
+            stop.position -= repeat_length * static_cast<float>(repeat_count);
+            if (stop.position < 0) {
+                stop.color = get_color_between_stops(0.0f, stop, color_stop_list_with_expanded_repeat.first());
+                stop.position = 0.0f;
+                color_stop_list_with_expanded_repeat.prepend(stop);
+                break;
+            }
+            color_stop_list_with_expanded_repeat.prepend(stop);
+        }
+    }
+
+    for (auto repeat_count = 1; repeat_count < positive_repeat_count; repeat_count++) {
+        for (auto stop : color_stop_list) {
+            stop.position += repeat_length * static_cast<float>(repeat_count);
+            if (stop.position > 1) {
+                stop.color = get_color_between_stops(1.0f, stop, color_stop_list_with_expanded_repeat.last());
+                stop.position = 1.0f;
+                color_stop_list_with_expanded_repeat.append(stop);
+                break;
+            }
+            color_stop_list_with_expanded_repeat.append(stop);
+        }
+    }
+
+    return color_stop_list_with_expanded_repeat;
+}
+
+static ColorStopList expand_color_stops_for_painting(ColorStopList const& color_stop_list, Optional<float> repeat_length)
+{
+    auto expanded = repeat_length.has_value()
+        ? expand_repeat_length(color_stop_list, repeat_length.value())
+        : color_stop_list;
+    return replace_transition_hints_with_normal_color_stops(expanded);
+}
 
 static ColorStopData resolve_color_stop_positions(Layout::NodeWithStyle const& node, Vector<CSS::ColorStopListElement> const& color_stop_list, auto resolve_position_to_float, bool repeating)
 {
@@ -112,7 +217,7 @@ static ColorStopData resolve_color_stop_positions(Layout::NodeWithStyle const& n
     if (repeating)
         repeat_length = resolved_color_stops.last().position - resolved_color_stops.first().position;
 
-    return { resolved_color_stops, repeat_length };
+    return { resolved_color_stops, repeat_length, repeating };
 }
 
 LinearGradientData resolve_linear_gradient_data(Layout::NodeWithStyle const& node, CSSPixelSize gradient_size, CSS::LinearGradientStyleValue const& linear_gradient)
@@ -133,6 +238,9 @@ LinearGradientData resolve_linear_gradient_data(Layout::NodeWithStyle const& nod
         },
         linear_gradient.is_repeating());
 
+    // Replace transition hints for painting; keep repeat_length for Skia's native tiling
+    resolved_color_stops.list = replace_transition_hints_with_normal_color_stops(resolved_color_stops.list);
+
     return { gradient_angle, resolved_color_stops, linear_gradient.interpolation_method() };
 }
 
@@ -144,6 +252,11 @@ ConicGradientData resolve_conic_gradient_data(Layout::NodeWithStyle const& node,
             return CSS::Angle::from_style_value(position, one_turn).to_degrees() / one_turn.to_degrees();
         },
         conic_gradient.is_repeating());
+
+    // Expand color stops for painting (replace transition hints and expand repeat length)
+    resolved_color_stops.list = expand_color_stops_for_painting(resolved_color_stops.list, resolved_color_stops.repeat_length);
+    resolved_color_stops.repeat_length = {};
+
     return { conic_gradient.angle_degrees(), resolved_color_stops, conic_gradient.interpolation_method() };
 }
 
@@ -163,6 +276,11 @@ RadialGradientData resolve_radial_gradient_data(Layout::NodeWithStyle const& nod
             return position.as_calculated().resolve_length(context)->absolute_length_to_px_without_rounding() / gradient_size.width().to_float();
         },
         radial_gradient.is_repeating());
+
+    // Expand color stops for painting (replace transition hints and expand repeat length)
+    resolved_color_stops.list = expand_color_stops_for_painting(resolved_color_stops.list, resolved_color_stops.repeat_length);
+    resolved_color_stops.repeat_length = {};
+
     return { resolved_color_stops, radial_gradient.interpolation_method() };
 }
